refactor: split query rpc surface
